+++ System.app/src/components/ui/input.tsx
@@ -5,7 +5,15 @@ export interface InputProps
   extends React.InputHTMLAttributes<HTMLInputElement> {}
 
 const Input = React.forwardRef<HTMLInputElement, InputProps>(
-  ({ className, type, ...props }, ref) => {
+  ({ className, type, onFocus, ...props }, ref) => {
+    // Auto-select content on focus for number inputs
+    const handleFocus = React.useCallback((e: React.FocusEvent<HTMLInputElement>) => {
+      if (type === 'number') {
+        e.target.select()
+      }
+      onFocus?.(e)
+    }, [type, onFocus])
+
     return (
       <input
         type={type}
@@ -14,6 +22,7 @@ const Input = React.forwardRef<HTMLInputElement, InputProps>(
           className
         )}
         ref={ref}
+        onFocus={handleFocus}
         {...props}
       />
     )